pyproject.toml
[project]
name = "othello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "maturin>=1.7.1",
    "python-fasthtml>=0.4.5",
]
dev-dependencies = [
    "ruff>=0.6.2",
]

[build-system]
requires = ["maturin>=1,<2"]
build-backend = "maturin"

[tool.uv]
managed = true
